+++ stroom-pipeline/src/main/java/stroom/pipeline/server/reader/InvalidXmlCharFilter.java
@@ -42,6 +42,16 @@ public class InvalidXmlCharFilter extends TransformReader {
     private final char[] buffer;
     private boolean hasReadAhead;
     private char readAhead;
+    private boolean eof;
+
+    public InvalidXmlCharFilter(final Reader in,
+                                final XmlChars xmlChars) {
+        super(in);
+        this.xmlChars = xmlChars;
+        this.replace = false;
+        this.replacementChar = ' ';
+        buffer = new char[StreamUtil.BUFFER_SIZE];
+    }
 
     public InvalidXmlCharFilter(final Reader in,
                                 final XmlChars xmlChars,
@@ -53,114 +63,6 @@ public class InvalidXmlCharFilter extends TransformReader {
         this.replacementChar = replacementChar;
         buffer = new char[StreamUtil.BUFFER_SIZE];
     }
-
-//    @Override
-//    public int read() throws IOException {
-//        if (hasReadAhead) {
-//            hasReadAhead = false;
-//            return readAhead;
-//        }
-//
-//        boolean ok = false;
-//        int val = -1;
-//        while (!ok) {
-//            val = in.read();
-//            if (val < 0) {
-//                return val;
-//            }
-//
-//            final char ch = (char) val;
-//            if (!xmlChars.isValid(ch)) {
-//                if (Character.isHighSurrogate(ch)) {
-//
-//                    final int val2 = in.read();
-//                    if (val2 < 0) {
-//                        modified = true;
-//                        return val2;
-//                    }
-//
-//                    final char ch2 = (char) val2;
-//                    if (Character.isLowSurrogate(ch2)) {
-//                        final int supplemental = Character.toCodePoint(ch, ch2);
-//                        if (xmlChars.isValid(supplemental)) {
-//                            readAhead = ch2;
-//                            hasReadAhead = true;
-//                            ok = true;
-//                        }
-//                    }
-//                }
-//
-//                if (!ok) {
-//                    modified = true;
-//                }
-//
-//            } else {
-//                ok = true;
-//            }
-//        }
-//
-//        return val;
-//    }
-//
-//    @Override
-//    public int read(final char[] cbuf, int off, int len) throws IOException {
-//        final int originalOff = off;
-//        if (hasReadAhead && len - off > 0) {
-//            hasReadAhead = false;
-//            cbuf[off++] = readAhead;
-//            --len;
-//        }
-//
-//        if (len > 0) {
-//            final int length = in.read(cbuf, off, len);
-//            if (length >= 0) {
-//                final int endOffset = off + length;
-//                while (off < endOffset) {
-//                    final char ch = cbuf[off];
-//                    if (!xmlChars.isValid(ch)) {
-//                        if (Character.isHighSurrogate(ch)) {
-//                            final int nextOff = off + 1;
-//                            final boolean usingBuffer = nextOff < endOffset;
-//
-//                            boolean validSurrogate = false;
-//
-//                            if (usingBuffer) {
-//                                final char ch2 = cbuf[nextOff];
-//                                if (Character.isLowSurrogate(ch2)) {
-//                                    final int supplemental = Character.toCodePoint(ch, ch2);
-//                                    validSurrogate = xmlChars.isValid(supplemental);
-//                                }
-//                            } else {
-//                                validSurrogate = validLowSurrogateAhead(ch);
-//                            }
-//
-//                            if (!validSurrogate) {
-//                                if (usingBuffer) {
-//                                    cbuf[off++] = REPLACEMENT_CHAR;
-//                                    cbuf[off] = REPLACEMENT_CHAR;
-//                                } else {
-//                                    cbuf[off] = REPLACEMENT_CHAR;
-//                                    readAhead = REPLACEMENT_CHAR;
-//                                }
-//                                modified = true;
-//                            }
-//
-//                        } else {
-//                            cbuf[off] = REPLACEMENT_CHAR;
-//                            modified = true;
-//                        }
-//                    }
-//                    off++;
-//                }
-//            }
-//        }
-//        return (originalOff == off) ? -1 : off - originalOff;
-//    }
-
-
-
-
-
 
     public int read() throws IOException {
         final char[] cb = new char[1];
@@ -189,77 +91,84 @@ public class InvalidXmlCharFilter extends TransformReader {
             --len;
 
             // If we've no more characters requested then just return 1 for the char we added.
-            if (len == 0) {
+            if (len == 0 || eof) {
                 return 1;
             }
+        } else if (eof) {
+            return -1;
         }
 
-        final char[] inBuffer = buffer;
+        final int maxLen = Math.min(len, buffer.length);
+        final int length = in.read(buffer, 0, maxLen);
+        eof = length < 0;
 
-        final int maxLen = Math.min(len, inBuffer.length);
-        final int length = in.read(inBuffer, 0, maxLen);
-        if (length >= 0) {
-            for (int i = 0; i < length; i++) {
-                final char ch = inBuffer[i];
+        for (int i = 0; i < length; i++) {
+            final char ch = buffer[i];
 
-                if (!xmlChars.isValid(ch)) {
-                    if (Character.isHighSurrogate(ch)) {
-                        final int nextOff = i + 1;
-                        final boolean usingBuffer = nextOff < length;
+            if (!xmlChars.isValid(ch)) {
+                if (Character.isHighSurrogate(ch)) {
+                    boolean validSurrogate = false;
 
-                        boolean validSurrogate = false;
-
-                        if (usingBuffer) {
-                            final char ch2 = inBuffer[nextOff];
-                            if (Character.isLowSurrogate(ch2)) {
-                                final int supplemental = Character.toCodePoint(ch, ch2);
-                                validSurrogate = xmlChars.isValid(supplemental);
-                            }
-                        } else {
-                            validSurrogate = validLowSurrogateAhead(ch);
+                    // Move the index on one.
+                    i++;
+                    // See if we can use the buffer to read the next char into.
+                    final boolean usingBuffer = i < length;
+                    if (usingBuffer) {
+                        final char ch2 = buffer[i];
+                        if (Character.isLowSurrogate(ch2)) {
+                            final int supplemental = Character.toCodePoint(ch, ch2);
+                            validSurrogate = xmlChars.isValid(supplemental);
                         }
-
-                        if (!validSurrogate) {
-                            if (replace) {
-                                cbuf[off++] = replacementChar;
-                                if (usingBuffer) {
-                                    cbuf[off++] = replacementChar;
-                                } else {
-                                    readAhead = replacementChar;
-                                }
-                            } else {
-                                hasReadAhead = false;
-                            }
-
-                            modified = true;
-                        } else {
-                            cbuf[off++] = ch;
-                        }
-
                     } else {
+                        validSurrogate = validLowSurrogateAhead(ch);
+                    }
+
+                    if (!validSurrogate) {
                         if (replace) {
                             cbuf[off++] = replacementChar;
+                            if (usingBuffer) {
+                                cbuf[off++] = replacementChar;
+                            } else {
+                                readAhead = replacementChar;
+                            }
+                        } else {
+                            hasReadAhead = false;
                         }
+
                         modified = true;
+                    } else {
+                        cbuf[off++] = ch;
+                        if (usingBuffer) {
+                            cbuf[off++] = buffer[i];
+                        }
                     }
+
                 } else {
-                    cbuf[off++] = ch;
+                    if (replace) {
+                        cbuf[off++] = replacementChar;
+                    }
+                    modified = true;
                 }
+            } else {
+                cbuf[off++] = ch;
             }
         }
+
         return (originalOff == off) ? -1 : off - originalOff;
     }
 
     private boolean validLowSurrogateAhead(final char ch) throws IOException {
-        final int val2 = in.read();
-        if (val2 >= 0) {
-            final char ch2 = (char) val2;
-            readAhead = ch2;
-            hasReadAhead = true;
+        if (!eof) {
+            final int val2 = in.read();
+            if (val2 >= 0) {
+                final char ch2 = (char) val2;
+                readAhead = ch2;
+                hasReadAhead = true;
 
-            if (Character.isLowSurrogate(ch2)) {
-                final int supplemental = Character.toCodePoint(ch, ch2);
-                return xmlChars.isValid(supplemental);
+                if (Character.isLowSurrogate(ch2)) {
+                    final int supplemental = Character.toCodePoint(ch, ch2);
+                    return xmlChars.isValid(supplemental);
+                }
             }
         }
         return false;